fix(skilavottord):  Move mileage field to deregister page and remove dead code (#13244)
* Move mileage field to deregister page and remove dead code

* Added debug logger

* Update logging

---------
