--- apps/skilavottord/web/pages/deregister-vehicle-km/[id].tsx
@@ -1,6 +1,0 @@
-import { Screen } from '@island.is/skilavottord-web/types'
-import { withLocale } from '@island.is/skilavottord-web/i18n'
-import { DeregisterConfirmKm } from '@island.is/skilavottord-web/screens'
-import { withApollo } from '@island.is/skilavottord-web/graphql/withApollo'
-
-export default withApollo(withLocale('is')(DeregisterConfirmKm as Screen))
