--- apps/skilavottord/web/pages/en/deregister-vehicle-km/index.ts
@@ -1,5 +1,0 @@
-import { Screen } from '@island.is/skilavottord-web/types'
-import { withLocale } from '@island.is/skilavottord-web/i18n'
-import { DeregisterOverviewKm } from '@island.is/skilavottord-web/screens'
-
-export default withLocale('en')(DeregisterOverviewKm as Screen)
