--- apps/skilavottord/web/pages/en/deregister-vehicle-km/select.tsx
@@ -1,6 +1,0 @@
-import { Screen } from '@island.is/skilavottord-web/types'
-import { withLocale } from '@island.is/skilavottord-web/i18n'
-import { DeregisterSelectKm } from '@island.is/skilavottord-web/screens'
-import { withApollo } from '@island.is/skilavottord-web/graphql/withApollo'
-
-export default withApollo(withLocale('en')(DeregisterSelectKm as Screen))
