--- apps/skilavottord/web/screens/DeregisterVehicleKm/Overview/components/CarsTable/index.ts
@@ -1,1 +1,0 @@
-export * from './CarsTable'
