--- apps/skilavottord/web/screens/DeregisterVehicleKm/Overview/index.tsx
@@ -1,1 +1,0 @@
-export { default as Overview } from './Overview'
