--- apps/skilavottord/web/screens/DeregisterVehicleKm/index.tsx
@@ -1,3 +1,0 @@
-export { default as ConfirmKM } from './Confirm/ConfirmKM'
-export { default as Select } from './Select'
-export * from './Overview'
